src/requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.0.6
django-cors-headers==2.2.0
django-phone-field==1.8.0
djangorestframework==3.8.2
idna==2.9
pkg-resources==0.0.0
pytz==2018.4
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.8
zope.interface==5.0.1